segment id is now 64 bit and only compiling latest exercises

+++ src/common/Config.hpp
@@ -6,7 +6,7 @@ namespace dbi {
 
    using TId = uint64_t; // 16 Bit most significant bits are used for record id
    using PageId = uint64_t; // 16 Bit most significant bits should be 0 otherwise page id is invalid
-   using SegmentId = uint32_t;
+   using SegmentId = uint64_t;
    using RecordId = uint16_t;
 
    inline PageId toPageId(TId id) {return id>>16;}
